requirements.txt
streamlit
rembg[cpu]
pillow
onnxruntime
